+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog - Agent Framework Toolkit
 
-## Version 1.0.0-rc2 (4th of March 2026)
+## Unreleased
+- Anthropic: Added Support for Structured Output (custom implementation as Anthropic doe not adhere to common practice, but outcome is the same).
+- Update the various Model Constants with the latest releases
+- AzureOpenAI: Endpoint Pattern `https://<name>.openai.azure.com/openai/v1` is now also auto-corrected by default (to: `https://<name>.openai.azure.com`)
+- Added `AgentSession.GetMessages()` extension method for easier discoverability on getting messages from a session (Require that session belong to an agent that use the default InMemoryChatHistory)
+
+---
+
+## Version 1.0.0-rc3 (4th of March 2026)
 - Bump Agent Framework from rc2 to rc3
 - Bump Google.GenAI NuGet from 1.1.0 to 1.2.0
 - Bump Azure.Identity from 1.17.1 to 1.18.0